Block touch events while page is expanding/collapsing

+++ inboxrecyclerview/src/main/java/me/saket/inboxrecyclerview/page/PullToCollapseListener.kt
@@ -15,13 +15,13 @@ class PullToCollapseListener(private val expandablePage: ExpandablePageLayout) :
   constructor(context: Context, expandablePage: ExpandablePageLayout) : this(expandablePage)
 
   /** Minimum Y-distance the page has to be pulled before it's eligible for collapse. */
-  var collapseDistanceThreshold: Int = (Views.toolbarHeight(expandablePage.context)).toInt()
+  var collapseDistanceThreshold: Int = (Views.toolbarHeight(expandablePage.context))
 
   private val touchSlop: Int = ViewConfiguration.get(expandablePage.context).scaledTouchSlop
   private val onPullListeners = ArrayList<OnPullListener>(3)
-  private var downX: Float = 0F
-  private var downY: Float = 0F
-  private var lastMoveY: Float = 0F
+  private var downX: Float = 0f
+  private var downY: Float = 0f
+  private var lastMoveY: Float = 0f
   private var eligibleForCollapse: Boolean = false
   private var horizontalSwipingConfirmed: Boolean? = null
   private var interceptedUntilNextGesture: Boolean? = null
